requirements.txt
streamlit==1.29.0
pandas>=1.3.0
scikit-learn
altair>=4.0
google-generativeai
python-dotenv
streamlit_lottie